fixing incorrect pulls
