+++ OverExposed-Journal/lib/journal_entry.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:transparent_image/transparent_image.dart';
+
+class JournalEntry {
+  const JournalEntry({
+    required this.path,
+    required this.date,
+    required this.title,
+    required this.description
+  });
+
+  final String path;
+  final DateTime date;
+  final String title;
+  final String description;
+
+  MemoryImage getImage() {
+    return MemoryImage(kTransparentImage);
+  }
+}
